requirements.txt
openai>=1.14.0
chromadb>=0.4.24
python-dotenv>=1.0.1
tiktoken>=0.6.0
pytest>=8.1.1
rich>=13.7.1
langchain-text-splitters>=0.0.1
